requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.1
black==19.10b0
chardet==3.0.4
chrome-handler @ git+https://github.com/kaito1002/chrome_handler@107b54730b649a312b648572ee80eefd2a5919fd
click==7.1.2
flake8==3.8.2
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.13.0
regex==2020.5.14
selenium==3.141.0
six==1.15.0
slackclient==2.6.2
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
